Changed data type for 5

+++ 0x13-bit_manipulation/5-flip_bits.c
@@ -9,8 +9,8 @@
   */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int flips = n ^ m;
-	unsigned int count = 0;
+	unsigned long int flips = n ^ m;
+	unsigned long int count = 0;
 
 	while (flips)
 	{
